pyproject.toml
[project]
name = "encoderops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.31.4",
    "ipykernel>=6.29.5",
    "kooplearn>=1.1.3",
    "linear-operator-learning>=0.2.3",
    "lmdb>=1.6.2",
    "loguru>=0.7.3",
    "matscipy>=1.1.1",
    "mdtraj>=1.10.3",
    "mlcolvar",
    "nglview>=3.1.4",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "setuptools>=78.1.0",
    "torch-geometric>=2.6.1",
    "torch-scatter>=2.1.2",
    "tqdm>=4.67.1",
    "treescope>=0.1.9",
    "tyro>=0.9.18",
    "wandb>=0.19.8",
    "xarray>=2025.4.0",
]

[tool.uv.sources]
mlcolvar = { git = "https://github.com/luigibonati/mlcolvar.git", rev = "graphs_integration" }
